backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.18
aiofiles>=24.0.0

# Audio processing
praat-parselmouth>=0.4.5
numpy>=1.26.0
scipy>=1.14.0
soundfile>=0.12.0
librosa>=0.10.0

# Utilities
pyyaml>=6.0.0
regex>=2024.0.0
